requirements.txt
openai
pytorch
numpy
pyaudio
requests
whisper
returns
python-dotenv
wave
gtts
playsound